Implement back-up fetching feature for the RB & Add tests

+++ app/src/main/java/com/github/factotum_sdp/factotum/ui/settings/SettingsFragment.kt
@@ -1,5 +1,6 @@
 package com.github.factotum_sdp.factotum.ui.settings
 
+import android.app.AlertDialog
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -9,6 +10,7 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import com.github.factotum_sdp.factotum.R
 import com.github.factotum_sdp.factotum.ui.roadbook.RoadBookViewModel
+import com.google.android.material.snackbar.Snackbar
 
 /**
  * The fragment holding the Application settings UI
@@ -17,6 +19,7 @@ class SettingsFragment: Fragment() {
 
     private val settings: SettingsViewModel by activityViewModels()
     private val roadBookViewModel: RoadBookViewModel by activityViewModels()
+
     private lateinit var roadBookPreferences: CheckBox
     private lateinit var loadRoadBookBackUp: CheckBox
     private lateinit var deleteAllRoadBook: CheckBox
@@ -28,23 +31,75 @@ class SettingsFragment: Fragment() {
         val view = inflater.inflate(R.layout.fragment_settings, container, false)
 
         roadBookPreferences = view.findViewById(R.id.save_roadbook_preferences)
-        roadBookPreferences.isChecked = settings.settingsLiveData.value?.useRoadBookPreferences ?: false
-        roadBookPreferences.setOnCheckedChangeListener { _, isChecked ->
-            settings.updateUseRoadBookPreferences(isChecked)
-        }
-
         loadRoadBookBackUp = view.findViewById(R.id.load_roadbook_backup)
-        loadRoadBookBackUp.isChecked = false
-        loadRoadBookBackUp.setOnCheckedChangeListener { _, isChecked ->
-            if (isChecked) roadBookViewModel.fetchBackBackUps()
-        }
-
         deleteAllRoadBook = view.findViewById(R.id.delete_all_roadbook)
-        deleteAllRoadBook.isChecked = false
-        deleteAllRoadBook.setOnCheckedChangeListener { _, isChecked ->
-            if(isChecked) roadBookViewModel.clearAllRecords()
-        }
+
+        setSettingItems()
 
         return view
+    }
+
+    override fun onResume() {
+        super.onResume()
+        loadRoadBookBackUp.isChecked = false
+        deleteAllRoadBook.isChecked = false
+    }
+
+    private fun setSettingItems() {
+        setCheckedSettingBehavior(
+            roadBookPreferences,
+            settings.settingsLiveData.value?.useRoadBookPreferences ?: false) {
+            settings.updateUseRoadBookPreferences(it)
+        }
+        setCheckedSettingBehavior(loadRoadBookBackUp, false) {
+            if(it) {
+                roadBookViewModel.fetchBackBackUps()
+                snapMessage(R.string.snap_load_roadbook_backup)
+            }
+        }
+        setCheckedSettingBehavior(deleteAllRoadBook, false) {
+            if(it) {
+                deleteAllConfirmationDialog({
+                    roadBookViewModel.clearAllRecords()
+                    snapMessage(R.string.snap_delete_all_roadbook)
+                    deleteAllRoadBook.isChecked = false
+                }, {
+                    deleteAllRoadBook.isChecked = false
+                })
+            }
+        }
+    }
+
+    private fun setCheckedSettingBehavior(
+        checkBox: CheckBox,
+        checkedInitState: Boolean,
+        onCheckedChange: (Boolean) -> Unit
+    )
+    {
+        checkBox.isChecked = checkedInitState
+        checkBox.setOnCheckedChangeListener{ _, isChecked ->
+            onCheckedChange(isChecked)
+        }
+    }
+
+    private fun deleteAllConfirmationDialog(onConfirm: () -> Unit, onCancel:() -> Unit) {
+        val builder = AlertDialog.Builder(requireContext())
+        builder.setTitle(getString(R.string.delete_all_roadbook_dialog_title))
+        builder.setCancelable(false)
+        builder.setPositiveButton(getString(R.string.positive_label_delete_all_roadbook_dialog)) { _, _ ->
+            onConfirm()
+        }
+        builder.setNegativeButton(getString(R.string.negative_label_delete_all_roadbook_dialog)){ _, _ ->
+            onCancel()
+        }
+        val dial = builder.create()
+        dial.window?.attributes?.windowAnimations = R.style.DialogAnimLeftToRight
+        dial.show()
+    }
+
+    private fun snapMessage(snapMessageID: Int) {
+        Snackbar
+            .make(requireView(), getString(snapMessageID), 1000)
+            .setAction("Action", null).show()
     }
 }
